wallet: make useTimeSince a generic interface function
Extracts the dynamic time since functionality from wallet logs
so that this can be reused.
